convert to getCustomNormalizedFilePath function
