+++ packages/babel-plugin-ember-test-metadata/__tests__/get-normalized-file-path-test.js
@@ -1,3 +1,4 @@
+const { basename, join, resolve, sep } = require('path');
 const { getNormalizedFilePath } = require('../get-normalized-file-path');
 
 describe('getNormalizedFilePath', () => {
@@ -159,8 +160,25 @@ describe('getNormalizedFilePath', () => {
     });
   });
 
-  describe('build isolation', () => {
+  describe('custom normalized filePath function', () => {
     const appRoot = '/Users/tester/workspace/personal/test-bed/example-app';
+
+    function normalizeBuildIsolationPath(filePath, packageName, projectRoot) {
+        const pathSegments = filePath.split(sep);
+        const testFilePath = pathSegments
+          .slice(pathSegments.indexOf('tests'))
+          .join(sep);
+      
+        if (!projectRoot) {
+          return testFilePath;
+        }
+
+        const addonRoot = pathSegments.slice(0, pathSegments.lastIndexOf(packageName));
+        const projectBaseDir = basename(resolve(addonRoot.join(sep), projectRoot));
+        const projectPath = addonRoot.slice(addonRoot.indexOf(projectBaseDir) + 1).join(sep);
+        
+        return join(projectPath, testFilePath);
+      }
 
     it('returns the normalized filepath for build isolation addon test', () => {
       const filePath = `${appRoot}/packages/addons/example-addon/example-app/tests/acceptance/foo-test.js`;
@@ -170,7 +188,7 @@ describe('getNormalizedFilePath', () => {
         root: appRoot,
         packageName: 'example-app',
         projectRoot: '../../..',
-        isBuildIsolation: true,
+        getCustomNormalizedFilePath: normalizeBuildIsolationPath,
       };
 
       const normalizedFilePath = getNormalizedFilePath(opts);
